data lazy loading isn't a good idea.
If data sources are loaded lazy, terraform will always ask for a change,
even there is none, as it doesn't know the data objects.
Defining another variable tag_ids and add them directly to the vm
instead of using a data source is the better way, if the tags are
created with terraform.